carbon-iot-complete/device-service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
